Code refactoring EVM-5
