+++ user-service/middleware/errorHandler.js
@@ -1,0 +1,6 @@
+const errorHandler = (err, req, res, next) => {
+    console.error(err.stack);
+    res.status(500).json({ message: 'Что-то пошло не так!' });
+};
+
+module.exports = errorHandler;
