+++ user-service/routes/userRoutes.js
@@ -1,0 +1,13 @@
+const express = require('express');
+const UserController = require('../controllers/userController');
+
+const router = express.Router();
+
+router.post('/', UserController.createUser);
+router.get('/', UserController.getAllUsers);
+router.get('/:id', UserController.getUserById);
+router.put('/:id', UserController.updateUser);
+router.delete('/:id', UserController.deleteUser);
+router.post('/login', UserController.login);
+
+module.exports = router;
